Refactor `metrics` and update README.md
