+++ commands/metrics/thermodynamic/thermodynamic.go
@@ -46,8 +46,9 @@ var Command = cli.Command{
 		end := ctx.String("end")
 		start := ctx.String("start")
 		step := ctx.Generic("step")
+
 		metricsName := ctx.String("name")
-		scope := ctx.Generic("scope").(*model.ScopeEnumValue).Selected
+		scope := interceptor.ParseScope(metricsName)
 
 		duration := schema.Duration{
 			Start: start,
